+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Export and import all KurpoBot data with JSON. [#22](https://github.com/hpopp/kurpo-bot/pull/22)
+
+### Changed
+
+- Various internal refactoring and testing. [#21](https://github.com/hpopp/kurpo-bot/pull/21)
+
 ## [0.6.6] - 2025-06-22
 
 ### Changed